+++ event-subscription/src/main/java/uk/gov/justice/subscription/domain/subscriptiondescriptor/SubscriptionsDescriptor.java
@@ -8,15 +8,18 @@ public class SubscriptionsDescriptor {
     private final String specVersion;
     private final String service;
     private final String serviceComponent;
+    private final int prioritisation;
     private final List<Subscription> subscriptions;
 
     public SubscriptionsDescriptor(final String specVersion,
                                    final String service,
                                    final String serviceComponent,
+                                   final int prioritisation,
                                    final List<Subscription> subscriptions) {
         this.specVersion = specVersion;
         this.service = service;
         this.serviceComponent = serviceComponent;
+        this.prioritisation = prioritisation;
         this.subscriptions = subscriptions;
     }
 
@@ -32,23 +35,39 @@ public class SubscriptionsDescriptor {
         return serviceComponent;
     }
 
-    @Override
-    public boolean equals(Object object) {
-        if (this == object) return true;
-        if (object == null || getClass() != object.getClass()) return false;
-        SubscriptionsDescriptor that = (SubscriptionsDescriptor) object;
-        return Objects.equals(specVersion, that.specVersion) &&
-                Objects.equals(service, that.service) &&
-                Objects.equals(serviceComponent, that.serviceComponent);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(specVersion, service, serviceComponent);
+    public int getPrioritisation() {
+        return prioritisation;
     }
 
     public List<Subscription> getSubscriptions() {
         return subscriptions;
     }
 
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (!(o instanceof SubscriptionsDescriptor)) return false;
+        final SubscriptionsDescriptor that = (SubscriptionsDescriptor) o;
+        return Objects.equals(specVersion, that.specVersion) &&
+                Objects.equals(service, that.service) &&
+                Objects.equals(serviceComponent, that.serviceComponent) &&
+                Objects.equals(prioritisation, that.prioritisation) &&
+                Objects.equals(subscriptions, that.subscriptions);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(specVersion, service, serviceComponent, prioritisation, subscriptions);
+    }
+
+    @Override
+    public String toString() {
+        return "SubscriptionsDescriptor{" +
+                "specVersion='" + specVersion + '\'' +
+                ", service='" + service + '\'' +
+                ", serviceComponent='" + serviceComponent + '\'' +
+                ", prioritisation='" + prioritisation + '\'' +
+                ", subscriptions=" + subscriptions +
+                '}';
+    }
 }